se crea el map y se consume info de la bd y se muestra en la card
